Don't fetch Construction Certificates
This changes the data we're getting from council to exclude Construction Certificates. Normally we want as much data as possible but before this change people were seeing what looked like duplicate alerts. This is because this council seems to create a Construction Certificates record at the same time as their Development Application record. These look pretty much identical to PlanningAlerts subscribers. The more important one is the DA because you can usually comment on it.

Fixes #5.